@@ -15,19 +15,16 @@ import org.springframework.web.bind.annotation.RestController;
 import com.ruoyi.common.core.controller.BaseController;
 import com.ruoyi.common.core.domain.R;
 import com.ruoyi.common.utils.StringUtils;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiImplicitParams;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.tags.Tag;
 
 /**
  * swagger 用户测试方法
- * 
+ *
  * @author ruoyi
  */
-@Api("用户信息管理")
+@Tag(name = "用户信息管理")
 @RestController
 @RequestMapping("/test/user")
 public class TestController extends BaseController
@@ -37,19 +34,19 @@ public class TestController extends BaseController
         users.put(1, new UserEntity(1, "admin", "admin123", "15888888888"));
         users.put(2, new UserEntity(2, "ry", "admin123", "15666666666"));
     }
-
-    @ApiOperation("获取用户列表")
+    
+    @Operation(summary = "获取用户列表")
     @GetMapping("/list")
     public R<List<UserEntity>> userList()
     {
         List<UserEntity> userList = new ArrayList<UserEntity>(users.values());
         return R.ok(userList);
     }
-
-    @ApiOperation("获取用户详细")
-    @ApiImplicitParam(name = "userId", value = "用户ID", required = true, dataType = "int", paramType = "path", dataTypeClass = Integer.class)
+    
+    @Operation(summary = "获取用户详细")
     @GetMapping("/{userId}")
-    public R<UserEntity> getUser(@PathVariable Integer userId)
+    public R<UserEntity> getUser(@PathVariable(name = "userId")
+    Integer userId)
     {
         if (!users.isEmpty() && users.containsKey(userId))
         {
@@ -60,14 +57,8 @@ public class TestController extends BaseController
             return R.fail("用户不存在");
         }
     }
-
-    @ApiOperation("新增用户")
-    @ApiImplicitParams({
-        @ApiImplicitParam(name = "userId", value = "用户id", dataType = "Integer", dataTypeClass = Integer.class),
-        @ApiImplicitParam(name = "username", value = "用户名称", dataType = "String", dataTypeClass = String.class),
-        @ApiImplicitParam(name = "password", value = "用户密码", dataType = "String", dataTypeClass = String.class),
-        @ApiImplicitParam(name = "mobile", value = "用户手机", dataType = "String", dataTypeClass = String.class)
-    })
+    
+    @Operation(summary = "新增用户")
     @PostMapping("/save")
     public R<String> save(UserEntity user)
     {
@@ -78,10 +69,11 @@ public class TestController extends BaseController
         users.put(user.getUserId(), user);
         return R.ok();
     }
-
-    @ApiOperation("更新用户")
+    
+    @Operation(summary = "更新用户")
     @PutMapping("/update")
-    public R<String> update(@RequestBody UserEntity user)
+    public R<String> update(@RequestBody
+    UserEntity user)
     {
         if (StringUtils.isNull(user) || StringUtils.isNull(user.getUserId()))
         {
@@ -95,11 +87,11 @@ public class TestController extends BaseController
         users.put(user.getUserId(), user);
         return R.ok();
     }
-
-    @ApiOperation("删除用户信息")
-    @ApiImplicitParam(name = "userId", value = "用户ID", required = true, dataType = "int", paramType = "path", dataTypeClass = Integer.class)
+    
+    @Operation(summary = "删除用户信息")
     @DeleteMapping("/{userId}")
-    public R<String> delete(@PathVariable Integer userId)
+    public R<String> delete(@PathVariable(name = "userId")
+    Integer userId)
     {
         if (!users.isEmpty() && users.containsKey(userId))
         {
@@ -113,26 +105,26 @@ public class TestController extends BaseController
     }
 }
 
-@ApiModel(value = "UserEntity", description = "用户实体")
+@Schema(description = "用户实体")
 class UserEntity
 {
-    @ApiModelProperty("用户ID")
+    @Schema(title = "用户ID")
     private Integer userId;
-
-    @ApiModelProperty("用户名称")
+    
+    @Schema(title = "用户名称")
     private String username;
-
-    @ApiModelProperty("用户密码")
+    
+    @Schema(title = "用户密码")
     private String password;
-
-    @ApiModelProperty("用户手机")
+    
+    @Schema(title = "用户手机")
     private String mobile;
-
+    
     public UserEntity()
     {
-
+        
     }
-
+    
     public UserEntity(Integer userId, String username, String password, String mobile)
     {
         this.userId = userId;
@@ -140,42 +132,42 @@ class UserEntity
         this.password = password;
         this.mobile = mobile;
     }
-
+    
     public Integer getUserId()
     {
         return userId;
     }
-
+    
     public void setUserId(Integer userId)
     {
         this.userId = userId;
     }
-
+    
     public String getUsername()
     {
         return username;
     }
-
+    
     public void setUsername(String username)
     {
         this.username = username;
     }
-
+    
     public String getPassword()
     {
         return password;
     }
-
+    
     public void setPassword(String password)
     {
         this.password = password;
     }
-
+    
     public String getMobile()
     {
         return mobile;
     }
-
+    
     public void setMobile(String mobile)
     {
         this.mobile = mobile;
